libwasi: move sock into a separate file
